Added execution time function for parallel codes

+++ oscpara.h
@@ -12,7 +12,7 @@ _simulate_damped_os_parallel_mpi_omp(double max_amplitude, double length, double
                                      double time_limit, double step_size, double damping_coefficent,
                                      int number_of_files);
 
-extern double
+extern int
 _execution_time_damped_os_parallel_mpi_omp(double max_amplitude, double length, double mass, double gravity, double k,
                                            double Ao,
                                            double Vo, double FI,
@@ -24,7 +24,7 @@ _simulate_damped_os_parallel_mpi(double max_amplitude, double length, double mas
                                  double Vo, double FI,
                                  double time_limit, double step_size, double damping_coefficent, int number_of_files);
 
-extern double
+extern int
 _execution_time_damped_os_parallel_mpi(double max_amplitude, double length, double mass, double gravity, double k,
                                        double Ao,
                                        double Vo, double FI,
